requirements.txt
google-genai
google-cloud-storage
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
